Performance testing (#13)
Uses JMeter to add load testing and stress testing for the SonarQube Web
API.

Load testing checks the performance of GET endpoints related to system
and project information.

Stress testing checks the performance of POST endpoints for creating and
deleting projects.

Moved the internal `dummy-project` to a more accessible location in the
`courseProjectCode` directory.

---------

Co-authored-by: Christian Ashley <christianzashley@gmail.com>